Added state to UserEditClub and fixed bug
Added state in UserEditClub component, representing the club the user currently has selected. pressing the edit club button sends that clubs ID as a query to the edit club page.

Also fixed a bug in which clubs would not load for admin users, due to race condition on ready variable.