Implement basic remove club logic
The UserPage passes the user's profile to the UserClubList. The UserClubList only really needs the user's email from the profile. The UserClubList has a callback function passed to the UserClubCard. This callback function passes the _id of the club to be left back up to the UserClubList, which updates the ClubsCollection to remove the user's email from the members array of that club.