Implementation of Join/Leave Club on BrowseClubs
Clubs state is stored in BrowseClubsPage. If clubs status changes, calls to re-render. ClubCard conditionally renders depending on whether the user is an Admin or not, and whether the user can edit the club. Also the Join or leave club button updates dynamically to reflect whether the user is a member of that club.